Refactor riskyHosts endpoint to support user risk score
