+++ x-pack/plugins/security_solution/common/search_strategy/security_solution/hosts/kpi/index.ts
@@ -9,7 +9,6 @@ export * from './authentications';
 export * from './common';
 export * from './hosts';
 export * from './unique_ips';
-export * from './risky_hosts';
 
 import { HostsKpiAuthenticationsStrategyResponse } from './authentications';
 import { HostsKpiHostsStrategyResponse } from './hosts';
@@ -21,7 +20,6 @@ export enum HostsKpiQueries {
   kpiHosts = 'hostsKpiHosts',
   kpiHostsEntities = 'hostsKpiHostsEntities',
   kpiUniqueIps = 'hostsKpiUniqueIps',
-  kpiRiskyHosts = 'hostsKpiRiskyHosts',
   kpiUniqueIpsEntities = 'hostsKpiUniqueIpsEntities',
 }
