+++ x-pack/plugins/security_solution/common/search_strategy/security_solution/risk_score/common/index.test.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { getHostRiskIndex, getUserRiskIndex } from '.';
+import { getHostRiskIndex, getUserRiskIndex } from './';
 
 describe('hosts risk search_strategy getHostRiskIndex', () => {
   it('should properly return host index if space is specified', () => {
