+++ x-pack/plugins/security_solution/public/hosts/components/kpi_hosts/risky_hosts/index.tsx
@@ -25,8 +25,6 @@ import {
 import { HostsKpiBaseComponentLoader } from '../common';
 import * as i18n from './translations';
 
-import { HostsKpiRiskyHostsStrategyResponse } from '../../../../../common/search_strategy/security_solution/hosts/kpi/risky_hosts';
-
 import { useInspectQuery } from '../../../../common/hooks/use_inspect_query';
 import { useErrorToast } from '../../../../common/hooks/use_error_toast';
 
@@ -35,7 +33,7 @@ import {
   HOST_RISK_INFO_BUTTON_CLASS,
 } from '../../host_risk_information';
 import { HoverVisibilityContainer } from '../../../../common/components/hover_visibility_container';
-import { RiskSeverity } from '../../../../../common/search_strategy';
+import { KpiRiskScoreStrategyResponse, RiskSeverity } from '../../../../../common/search_strategy';
 import { RiskScore } from '../../../../common/components/severity/common';
 
 const QUERY_ID = 'hostsKpiRiskyHostsQuery';
@@ -62,7 +60,7 @@ const RiskScoreContainer = styled(EuiFlexItem)`
 const RiskyHostsComponent: React.FC<{
   error: unknown;
   loading: boolean;
-  data?: HostsKpiRiskyHostsStrategyResponse;
+  data?: KpiRiskScoreStrategyResponse;
 }> = ({ error, loading, data }) => {
   useInspectQuery(QUERY_ID, loading, data);
   useErrorToast(i18n.ERROR_TITLE, error);
@@ -71,8 +69,8 @@ const RiskyHostsComponent: React.FC<{
     return <HostsKpiBaseComponentLoader />;
   }
 
-  const criticalRiskCount = data?.riskyHosts.Critical ?? 0;
-  const hightlRiskCount = data?.riskyHosts.High ?? 0;
+  const criticalRiskCount = data?.kpiRiskScore.Critical ?? 0;
+  const hightlRiskCount = data?.kpiRiskScore.High ?? 0;
 
   const totalCount = criticalRiskCount + hightlRiskCount;
