+++ x-pack/plugins/security_solution/public/hosts/containers/risk_score/all/translations.ts
