+++ x-pack/plugins/security_solution/public/overview/components/overview_risky_host_links/index.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 
 import { RiskyHostsEnabledModule } from './risky_hosts_enabled_module';
 import { RiskyHostsDisabledModule } from './risky_hosts_disabled_module';
-import { useHostRiskScore } from '../../../hosts/containers/host_risk_score';
+import { useHostRiskScore } from '../../../hosts/containers/risk_score/all';
 import { useQueryInspector } from '../../../common/components/page/manage_query';
 import { HostRiskScoreQueryId } from '../../../common/containers/hosts_risk/types';
 import { GlobalTimeArgs } from '../../../common/containers/use_global_time';
