+++ x-pack/plugins/security_solution/public/overview/pages/overview.test.tsx
@@ -27,7 +27,7 @@ import { useCtiDashboardLinks } from '../containers/overview_cti_links';
 import { useIsExperimentalFeatureEnabled } from '../../common/hooks/use_experimental_features';
 import { initialUserPrivilegesState } from '../../common/components/user_privileges/user_privileges_context';
 import { EndpointPrivileges } from '../../../common/endpoint/types';
-import { useHostRiskScore } from '../../hosts/containers/host_risk_score';
+import { useHostRiskScore } from '../../hosts/containers/risk_score/all';
 
 jest.mock('../../common/lib/kibana');
 jest.mock('../../common/containers/source');
