+++ x-pack/plugins/security_solution/public/timelines/components/side_panel/event_details/index.tsx
@@ -35,7 +35,7 @@ import { ALERT_DETAILS } from './translations';
 import { useWithCaseDetailsRefresh } from '../../../../common/components/endpoint/host_isolation/endpoint_host_isolation_cases_context';
 import { EventDetailsFooter } from './footer';
 import { EntityType } from '../../../../../../timelines/common';
-import { useHostRiskScore } from '../../../../hosts/containers/host_risk_score';
+import { useHostRiskScore } from '../../../../hosts/containers/risk_score/all';
 import { HostRisk } from '../../../../common/containers/hosts_risk/types';
 import { buildHostNamesFilter } from '../../../../../common/search_strategy';
