+++ x-pack/plugins/security_solution/server/search_strategy/security_solution/factory/hosts/index.ts
@@ -21,7 +21,6 @@ import { authentications, authenticationsEntities } from './authentications';
 import { hostsKpiAuthentications, hostsKpiAuthenticationsEntities } from './kpi/authentications';
 import { hostsKpiHosts, hostsKpiHostsEntities } from './kpi/hosts';
 import { hostsKpiUniqueIps, hostsKpiUniqueIpsEntities } from './kpi/unique_ips';
-import { hostsKpiRiskyHosts } from './kpi/risky_hosts';
 
 export const hostsFactory: Record<
   HostsQueries | HostsKpiQueries,
@@ -39,7 +38,6 @@ export const hostsFactory: Record<
   [HostsKpiQueries.kpiAuthenticationsEntities]: hostsKpiAuthenticationsEntities,
   [HostsKpiQueries.kpiHosts]: hostsKpiHosts,
   [HostsKpiQueries.kpiHostsEntities]: hostsKpiHostsEntities,
-  [HostsKpiQueries.kpiRiskyHosts]: hostsKpiRiskyHosts,
   [HostsKpiQueries.kpiUniqueIps]: hostsKpiUniqueIps,
   [HostsKpiQueries.kpiUniqueIpsEntities]: hostsKpiUniqueIpsEntities,
 };
