--- x-pack/plugins/security_solution/server/search_strategy/security_solution/factory/hosts/kpi/risky_hosts/index.ts
@@ -1,53 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { getOr } from 'lodash/fp';
-
-import type { IEsSearchResponse } from '../../../../../../../../../../src/plugins/data/common';
-import type { HostsKpiQueries } from '../../../../../../../common/search_strategy';
-import { RiskSeverity } from '../../../../../../../common/search_strategy';
-
-import type {
-  HostsKpiRiskyHostsRequestOptions,
-  HostsKpiRiskyHostsStrategyResponse,
-} from '../../../../../../../common/search_strategy/security_solution/hosts/kpi/risky_hosts';
-import { inspectStringifyObject } from '../../../../../../utils/build_query';
-import type { SecuritySolutionFactory } from '../../../types';
-import { buildHostsKpiRiskyHostsQuery } from './query.hosts_kpi_risky_hosts.dsl';
-
-interface AggBucket {
-  key: RiskSeverity;
-  doc_count: number;
-}
-
-export const hostsKpiRiskyHosts: SecuritySolutionFactory<HostsKpiQueries.kpiRiskyHosts> = {
-  buildDsl: (options: HostsKpiRiskyHostsRequestOptions) => buildHostsKpiRiskyHostsQuery(options),
-  parse: async (
-    options: HostsKpiRiskyHostsRequestOptions,
-    response: IEsSearchResponse<unknown>
-  ): Promise<HostsKpiRiskyHostsStrategyResponse> => {
-    const inspect = {
-      dsl: [inspectStringifyObject(buildHostsKpiRiskyHostsQuery(options))],
-    };
-
-    const riskBuckets = getOr([], 'aggregations.risk.buckets', response.rawResponse);
-
-    const riskyHosts: Record<RiskSeverity, number> = riskBuckets.reduce(
-      (cummulative: Record<string, number>, bucket: AggBucket) => ({
-        ...cummulative,
-        [bucket.key]: getOr(0, 'unique_hosts.value', bucket),
-      }),
-      {}
-    );
-
-    return {
-      ...response,
-      riskyHosts,
-      inspect,
-    };
-  },
-};
