+++ x-pack/plugins/security_solution/server/search_strategy/security_solution/factory/risk_score/all/query.risk_score.dsl.ts
@@ -11,9 +11,8 @@ import {
   RiskScoreRequestOptions,
   RiskScoreFields,
   RiskScoreSortField,
-} from '../../../../../common/search_strategy';
-
-import { createQueryFilterClauses } from '../../../../utils/build_query';
+} from '../../../../../../common/search_strategy';
+import { createQueryFilterClauses } from '../../../../../utils/build_query';
 
 export const QUERY_SIZE = 10;
